Change the way to calculate the solution

+++ include/Trick.hpp
@@ -7,9 +7,9 @@ class Trick {
 public:
     int32_t m_baseline_score;
     uint32_t m_time_trick;
-    uint8_t m_index;
+    uint16_t m_index;
 
-    Trick(int32_t baseline_score, uint32_t time_trick, uint8_t index);
+    Trick(int32_t baseline_score, uint32_t time_trick, uint16_t index);
 
     Trick(const Trick &t);
